+++ src/debug/runners/Linkserver.ts
@@ -1,4 +1,9 @@
+import fs from "fs";
 import { RunnerType, WestRunner } from "./WestRunner";
+import { compareVersions } from "../../utils";
+import path from "path";
+import { execPath } from "process";
+import { execCommandWithEnv } from "../../execUtils";
 
 export class Linkserver extends WestRunner {
   name = 'linkserver';
@@ -9,7 +14,11 @@ export class Linkserver extends WestRunner {
   get executable(): string | undefined{
     const exec = super.executable;
     if(!exec) {
-      return 'LinkServer';
+      if(process.platform === 'win32') {
+        return 'LinkServer.exe';
+      } else {
+        return 'LinkServer';
+      }
     }
   }
 
@@ -32,15 +41,116 @@ export class Linkserver extends WestRunner {
       }
     }
 
+    // Search in system
+    if(!this.serverPath || this.serverPath.length === 0 ) {
+      this.serverPath = this.findSystemLinkServer();
+    }
+
     if(args) {
       this.loadUserArgs(args);
     }
   }
 
+  async detect(): Promise<boolean> {
+    let found = await super.detect();
+
+    if(found) {
+      return true;
+    }
+
+    return new Promise<boolean>(async (resolve, reject) => {
+      let execPath = await this.findSystemLinkServer();
+      
+      if (!execPath) {
+        resolve(false);
+        return;
+      }
+
+      let versionCmd = `${execPath} --version`;
+      if(process.platform === 'linux' || process.platform === 'darwin') {
+        versionCmd = `${versionCmd} 2>&1`;
+      }
+      
+      execCommandWithEnv(`${versionCmd}`, (error: any, stdout: string, stderr: any) => {
+        if (error) {
+          resolve(false);
+        } else {
+          resolve(true);
+        }
+      });
+    });
+    
+  }
+
+  findSystemLinkServer(): string | undefined {
+    let directoryPath = '';
+    if(process.platform === 'win32') {
+      directoryPath = 'C:\\NXP';
+    } else {
+      directoryPath = '/usr/local';
+    }
+
+    const latestFile = this.findLatestVersion(directoryPath);
+
+    if (latestFile && this.executable) {
+      return path.join(latestFile, this.executable);
+    }
+
+    return undefined;
+  }
+
+  getVersionFromPath(filePath: string): string | null {
+    const match = filePath.match(/LinkServer_(\d+\.\d+\.\d+)/);
+    return match ? match[1] : null;
+  }
+
+  findLatestVersion(directory: string): string | null {
+    try {
+      const files = fs.readdirSync(directory);
+      
+      if (files.length === 0) {
+        return null;
+      }
+  
+      let latestVersion: string | null = null;
+      let latestFilePath: string | null = null;
+  
+      for (const file of files) {
+        const filePath = path.join(directory, file);
+        const version = this.getVersionFromPath(filePath);
+  
+        if (version) {
+          if (!latestVersion || compareVersions(version, latestVersion) > 0) {
+            latestVersion = version;
+            latestFilePath = filePath;
+          }
+        }
+      }
+  
+      return latestFilePath;
+    } catch (error) {
+      return null;
+    }
+  }
+  
   get autoArgs(): string {
     let cmdArgs = super.autoArgs;
+
     if(this.serverPath) {
       cmdArgs += ` --linkserver ${this.serverPath}`;
+    } else {
+
+      let pathExecSetting = this.getSetting('pathExec');
+      if(pathExecSetting) {
+        cmdArgs += ` --linkserver ${pathExecSetting}`;
+      } else {
+        // Search in system
+        let pathExec = this.findSystemLinkServer();
+        if(pathExec && pathExec.length > 0) {
+          cmdArgs += ` --linkserver ${pathExec}`;
+        }
+      }
+      
     }
     return cmdArgs;
   }